requirements.txt
requests>=2.14.0
PyYAML>=5.1.1
python-dateutil>=2.8.0
